Replace DrawingArea concept with CircuitDraw
Instead of thinking of drawing primitives as (monadic) functions on a
surface data structure, think of them as operations within a special
kind of monad (inspired by the Render monad in
Graphics.Rendering.Cairo). Without loss of generality, since one can
always use ReaderT to encapsulate the formerly explicit parameter.

This shifts a little bit of complexity from the CircuitDrawing module to
AsciiArtDrawing (which is still much simpler than the former), with no
relevant impact on CairoDrawing. Also, the code does not depend on the
non-standard language feature MultiParamTypeClasses any more.